refactor(shared): consolidate connector token encryption (#30125)
